Fix routes duplication
When a component declaring routes is re-rendered, the routes were
duplicated internally, leading to panic on refreshRoutes, because
vecty.Rerender panics if the component has been rendered.

The fix stores the routes in a map, and keep the oldest.

An other fix on refreshRoute invokes vecty.Rerender only on routes that
match the current pathname. This also prevents panic if a route has
never been rendered.